issue #8 remove leading spaces
Trim leading (and trailing) spaces & other whitespace from messages &
passwords to compensate for copy/paste operations that sometimes let in
spaces which in turn changes the alignment of the message rendering it
undecryptable.  This will make it line up consistently and allow greater
chance of decrypting when user has the correct password and message.